Finished everything except for subsystems.
ChiliInformer is made for streaming all the robot data via any way
(socket, websocket, writer, etc.) as a Map or a JSON object.

Beware with a random vision exception due to lack of memory, I couldn't
replicate the bug in the last couple of commits, but still.

ChiliIntake is missing code in the update() function with a ChiliPID
controller to regulate angle (ChiliPID2 is untested and not recommended
yet). It also need code to control the roller.

The hanger is uncertain, but the base class is there regardless.

To use a second camera, uncomment the code in ChiliVision and
SensorInput to enable control of the Axis cam and second C270. This
requires a custom Dashboard which must retain images after reception
until a new update (The CameraServer class changes the JPG file being
served too quickly and not retaining would mean blinking).

The server is located at roborio-2576-frc.local:5800 and includes a
table of Wiring information. The code needs an extra servlet class to
send the JSON located in ChiliInformer.

The DriverInput is missing the implementation of Informable and
publishing its data to ChiliInformer.

Auto recorder and playback not available.

Autonomous Maestro not available. Place autonomous code in
ChiliRobot.java. Remember to use a while loop to exit safely after 15
seconds.

ChiliDrive is missing code as well. Besided basic driving (trivial),
implement a first run conditional to determine the original heading of
the bot. Then, with the press of a button, run a P or PI controller to
return to said position or orient robot with the POV cap to 8 distinct
angles based on the original heading.

Implement custom driver station.

Good luck!

Lucas Alvarez.